Added tests to check if a registration in the app is being saved in the db. Currently it is failing

+++ src/main/java/no/steria/swhrs/HourRegDao.java
@@ -12,5 +12,5 @@ public interface HourRegDao {
 	
 	void saveHours(HourRegistration reg);
 	
-	List<HourRegistration> getHours(Long person_id, LocalDate date);
+	List<HourRegistration> getHours(int person_id, LocalDate date);
 }
